[red] support i32 as default generic constant value

+++ src/NFun.SyntaxTests/UserFunctions/ConcreteUserFunctionsTest.cs
@@ -32,7 +32,7 @@ namespace NFun.SyntaxTests.UserFunctions
         [TestCase("_inc(y) = y+1\r y = _inc(2)", 3)]
         [TestCase("mult2(a,b) = a*b \r y = mult2(3,4)+1", 13)]
         [TestCase("div2(a,b) = a/b  \r mult2(a,b) = a*b         \r y = mult2(3,4)+div2(4,2)", 14)]
-        public void ConstantEquation_NonRecursiveGenericFunction(string expr, double expected) => 
+        public void ConstantEquation_NonRecursiveGenericFunction(string expr, int expected) => 
             expr.AssertReturns("y",expected);
 
         [TestCase("_inc(a) = a+1.0\r y = _inc(2.0)",3.0)]
@@ -41,7 +41,7 @@ namespace NFun.SyntaxTests.UserFunctions
             expr.AssertReturns("y",expected);
 
         [TestCase("plus3(a,b,c) = plus2(a,b)+c \r plus2(a,b) = a+b  \r y = plus3(16,4,2)",22)]
-        public void ConstantEquation_ReversedImplementationsOfFunctions(string expr, double expected) => 
+        public void ConstantEquation_ReversedImplementationsOfFunctions(string expr, int expected) => 
             expr.AssertReturns("y",expected);
 
         [Test]
